Student validation added - Sagar

+++ src/main/kotlin/com/sagardhone/student/StudentServiceImpl.kt
@@ -1,33 +1,58 @@
 package com.sagardhone.library_management.student
 
+import exception.CustomException
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
+import java.util.Optional
 
 @Service
 class StudentServiceImpl(var studentRepository: StudentRepository):StudentService {
 
     override fun getAllStudentDetails(): List<Student> = studentRepository.findAll()
 
-    override fun getStudentDetailsById(studentId: String): Student = studentRepository.findById(studentId).orElseThrow({RuntimeException("Student not found.")})
+    override fun getStudentDetailsById(studentId: String): Student {
+
+        val foundStudent: Optional<Student> = studentRepository.findById(studentId)
+
+        if (foundStudent.isPresent()) return foundStudent.get() else throw CustomException(HttpStatus.NOT_ACCEPTABLE.value(),"Given student id $studentId not exist.!")
+    }
 
     override fun addStudentDetails(student: Student): Student = studentRepository.save(student)
 
     override fun updateStudentDetailsById(studentId: String, student: Student): Student {
-        val oldDetails =  studentRepository.findById(studentId).orElseThrow({RuntimeException("Student not found.")})
-        oldDetails.firstName = student.firstName
-        oldDetails.lastName = student.lastName
-        oldDetails.age = student.age
-        oldDetails.year = student.year
-        oldDetails.department = student.department
-        return studentRepository.save(oldDetails)
+        val foundStudent: Optional<Student> = studentRepository.findById(studentId)
+
+        if (foundStudent.isPresent){
+            val oldDetails =  foundStudent.get()
+            oldDetails.firstName = student.firstName
+            oldDetails.lastName = student.lastName
+            oldDetails.age = student.age
+            oldDetails.year = student.year
+            oldDetails.department = student.department
+            return studentRepository.save(oldDetails)
+        }else{
+            throw CustomException(HttpStatus.NOT_ACCEPTABLE.value(),"Given student id $studentId not exist.!")
+        }
+
     }
 
     override fun removeStudentDetailsById(studentId: String): String {
-        studentRepository.deleteById(studentId)
-        return "Student details removed for Student Id : $studentId"
+        val foundStudent: Optional<Student> = studentRepository.findById(studentId)
+
+        if (foundStudent.isPresent){
+            studentRepository.deleteById(studentId)
+            return "Student details removed for Student Id : $studentId"
+        }else{
+            throw CustomException(HttpStatus.NOT_ACCEPTABLE.value(),"Given student id $studentId not exist.!")
+        }
+
     }
 
     override fun removeMultipleStudentDetailsByIds(studentIds: List<String>): String {
-        studentRepository.deleteAllByIdInBatch(studentIds)
-        return "Student details removed for these Student Ids : $studentIds"
+        val foundStudent: List<String> = studentRepository.availableStudentIds(studentIds)
+
+        if (foundStudent.isEmpty()) throw CustomException(HttpStatus.NOT_FOUND.value(),"Given student ids $studentIds not exist.!")
+        studentRepository.deleteAllByIdInBatch(foundStudent)
+        return "Student details removed for these Book Ids : $foundStudent, \nthese ids not available ${studentIds.toSet() - foundStudent.toSet()}."
     }
 }
